RIVIERA-50: As I started integration testing, a number of changes to the validation of the JAR files emerged.  Of course all the validation still needs to be done, but it is being simplified.
One change is to remove the custom bean validation constraint on {{JarEntry}} to check to see if the version number of the SQL document starts with the version number of the directory.  This I moved as a check in the {{JarEntry}} constructor, and if they are not the same, a RuntimeException is thrown.  

Another change is the {{JarEntry}} no longer needs to store the version of the directory and the version of the SQL file separately...the {{JarEntry}} constructor checks to make sure they are the same and if they are, why store both?  In addition, there is no need to error check both separately, i.e., I no longer need a "directory version" check and a "file version" check.  The Major.Feature.Bug values of the two are the same so I only need to do a single check on each of Major.Feature.Bug.