RIVIERA-28: While doing testing on handling comments, I ran into a problem where the string literal in the SQL lookeed like a comment.  this is what it looked like
insert into foo (col1) values ('col1 value – Looks like a comment');

This didn't work because ' character is matched by the regex to match any character.  So the #replaceAll removed "-- Looks like a comment '" which is not correct.

This problem starts getting into the need for a full-featured SQL parser (maybe I can extract soemthing from Derby) and I don't want to go there.  It's not worth the time or effort. 

So i'm thinking I can solve 99% of the problems if I pattern match all characters except the ' character.  If I do this, the #replaceAll won't do the replacement in the above example anymore.  Of course this means that any comment that has the ' character won't be replaced anymore and will lead to exceptions then trying to execute the statement, but, I think it will be fine 99% of the time so I'm OK with that for now.